+++ src/lldb.ts
@@ -1,4 +1,4 @@
-import { MI2DebugSession, RunCommand } from './mibase';
+import { MI2DebugSession, RunCommand, SharedState } from './mibase';
 import { DebugSession, InitializedEvent, TerminatedEvent, StoppedEvent, OutputEvent, Thread, StackFrame, Scope, Source, Handles } from 'vscode-debugadapter';
 import { DebugProtocol } from 'vscode-debugprotocol';
 import { MI2_LLDB } from "./backend/mi2/mi2lldb";
@@ -39,6 +39,10 @@ export interface AttachRequestArguments extends DebugProtocol.AttachRequestArgum
 }
 
 class LLDBDebugSession extends MI2DebugSession {
+	constructor(debuggerLinesStartAt1?: boolean, isServer?: boolean) {
+		super(new SharedState(), debuggerLinesStartAt1, isServer)
+	}
+
 	protected override initializeRequest(response: DebugProtocol.InitializeResponse, args: DebugProtocol.InitializeRequestArguments): void {
 		response.body.supportsGotoTargetsRequest = true;
 		response.body.supportsHitConditionalBreakpoints = true;
@@ -55,20 +59,20 @@ class LLDBDebugSession extends MI2DebugSession {
 			this.sendErrorResponse(response, 104, `Configured debugger ${dbgCommand} not found.`);
 			return;
 		}
-		this.miDebugger = new MI2_LLDB(dbgCommand, [], args.debugger_args, args.env);
+		this.shared.miDebugger = new MI2_LLDB(dbgCommand, [], args.debugger_args, args.env);
 		this.setPathSubstitutions(args.pathSubstitutions);
 		this.initDebugger();
 		this.quit = false;
 		this.attached = false;
 		this.initialRunCommand = RunCommand.RUN;
-		this.isSSH = false;
+		this.shared.isSSH = false;
 		this.started = false;
 		this.crashed = false;
 		this.setValuesFormattingMode(args.valuesFormatting);
-		this.miDebugger.printCalls = !!args.printCalls;
-		this.miDebugger.debugOutput = !!args.showDevDebugOutput;
-		this.stopAtEntry = args.stopAtEntry;
-		this.miDebugger.registerLimit = args.registerLimit ?? "";
+		this.shared.miDebugger.printCalls = !!args.printCalls;
+		this.shared.miDebugger.debugOutput = !!args.showDevDebugOutput;
+		this.shared.stopAtEntry = args.stopAtEntry;
+		this.shared.miDebugger.registerLimit = args.registerLimit ?? "";
 		if (args.ssh !== undefined) {
 			if (args.ssh.forwardX11 === undefined)
 				args.ssh.forwardX11 = true;
@@ -80,15 +84,15 @@ class LLDBDebugSession extends MI2DebugSession {
 				args.ssh.x11host = "localhost";
 			if (args.ssh.remotex11screen === undefined)
 				args.ssh.remotex11screen = 0;
-			this.isSSH = true;
+			this.shared.isSSH = true;
 			this.setSourceFileMap(args.ssh.sourceFileMap, args.ssh.cwd, args.cwd);
-			this.miDebugger.ssh(args.ssh, args.ssh.cwd, args.target, args.arguments, undefined, false, args.autorun || []).then(() => {
+			this.shared.miDebugger.ssh(args.ssh, args.ssh.cwd, args.target, args.arguments, undefined, false, args.autorun || []).then(() => {
 				this.sendResponse(response);
 			}, err => {
 				this.sendErrorResponse(response, 106, `Failed to SSH: ${err.toString()}`);
 			});
 		} else {
-			this.miDebugger.load(args.cwd, args.target, args.arguments, undefined, args.autorun || []).then(() => {
+			this.shared.miDebugger.load(args.cwd, args.target, args.arguments, undefined, args.autorun || []).then(() => {
 				this.sendResponse(response);
 			}, err => {
 				this.sendErrorResponse(response, 107, `Failed to load MI Debugger: ${err.toString()}`);
@@ -102,19 +106,19 @@ class LLDBDebugSession extends MI2DebugSession {
 			this.sendErrorResponse(response, 104, `Configured debugger ${dbgCommand} not found.`);
 			return;
 		}
-		this.miDebugger = new MI2_LLDB(dbgCommand, [], args.debugger_args, args.env);
+		this.shared.miDebugger = new MI2_LLDB(dbgCommand, [], args.debugger_args, args.env);
 		this.setPathSubstitutions(args.pathSubstitutions);
 		this.initDebugger();
 		this.quit = false;
 		this.attached = true;
 		this.initialRunCommand = args.stopAtConnect ? RunCommand.NONE : RunCommand.CONTINUE;
-		this.isSSH = false;
+		this.shared.isSSH = false;
 		this.setValuesFormattingMode(args.valuesFormatting);
-		this.miDebugger.printCalls = !!args.printCalls;
-		this.miDebugger.debugOutput = !!args.showDevDebugOutput;
-		this.stopAtEntry = args.stopAtEntry;
-		this.miDebugger.registerLimit = args.registerLimit ?? "";
-		this.miDebugger.attach(args.cwd, args.executable, args.target, args.autorun || []).then(() => {
+		this.shared.miDebugger.printCalls = !!args.printCalls;
+		this.shared.miDebugger.debugOutput = !!args.showDevDebugOutput;
+		this.shared.stopAtEntry = args.stopAtEntry;
+		this.shared.miDebugger.registerLimit = args.registerLimit ?? "";
+		this.shared.miDebugger.attach(args.cwd, args.executable, args.target, args.autorun || []).then(() => {
 			this.sendResponse(response);
 		}, err => {
 			this.sendErrorResponse(response, 108, `Failed to attach: ${err.toString()}`);
@@ -125,7 +129,7 @@ class LLDBDebugSession extends MI2DebugSession {
 	protected setPathSubstitutions(substitutions: { [index: string]: string }): void {
 		if (substitutions) {
 			Object.keys(substitutions).forEach(source => {
-				this.miDebugger.extraCommands.push("settings append target.source-map " + source + " " + substitutions[source]);
+				this.shared.miDebugger.extraCommands.push("settings append target.source-map " + source + " " + substitutions[source]);
 			});
 		}
 	}
